pyproject.toml
[tool.poetry]
name = "personal-ai"
version = "0.1.0"
description = "Personal AI Assistant - Local-first AI with RAG and tools"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "app"},
    {include = "llm"},
    {include = "rag"},
    {include = "agents"},
    {include = "connectors"},
    {include = "core"},
    {include = "cli"},
]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
pydantic = ">=2.5.0"
pydantic-settings = ">=2.1.0"
httpx = ">=0.25.0"
# RAG and ML dependencies
sentence-transformers = ">=2.2.0"
qdrant-client = ">=1.7.0"
torch = ">=2.0.0"
transformers = ">=4.30.0"
# Additional utilities
python-dotenv = ">=1.0.0"
typer = {extras = ["all"], version = ">=0.9.0"}
# Document parsing
python-multipart = ">=0.0.6"
pypdf2 = ">=3.0.0"
python-docx = ">=1.1.0"
# Job queue
redis = ">=5.0.0"
arq = ">=0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-asyncio = ">=0.21"
ruff = "*"
black = "*"
mypy = ">=1.0.0"

[tool.poetry.scripts]
myai = "cli.main:main"
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120
select = [
    "A", "ARG", "B", "C", "DTZ", "E", "EM", "F", "FBT", "I", "ICN", "ISC", "N",
    "PLC", "PLE", "PLR", "PLW", "Q", "RUF", "S", "T", "TID", "UP", "W", "YTT"
]
ignore = [
    "B027", "FBT003", "S105", "S106", "S107", "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915"
]

[tool.black]
target-version = ["py311"]
line-length = 120
skip-string-normalization = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
